requirements.txt
transformers==4.48.0
accelerate==0.29.1
datasets==2.20.0
huggingface-hub==0.23.2
numpy==1.26.4
peft==0.11.1
pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
wandb==0.17.7
deepspeed==0.14.4
bitsandbytes==0.43.1
scipy
safetensors==0.4.3
scikit-learn
prettytable
pandas
ipython
flash-attn
timm
fire
opencv-python